Add function remove_noise_data to write atom data to a csv file
This commit adds a new function `remove_noise_data` to the `element.py` module in the `utils` directory. This function is responsible for writing atom data to a csv file. It takes `task_path`, `task_name`, `element`, `resolution`, and `final_out` as inputs and creates a csv file with the appropriate headers and data. The csv file is named based on the combination of `task_name` and `element`.