Add HDF5Loader class for loading data
This commit adds a new module, `data_tools.py`, which contains the `HDF5Loader` class. This class is responsible for loading data from an HDF5 file. It extends the `QThread` class and emits a `data_loaded` signal when the data has been loaded. The class also has a `set_task` method to set the folder name and full path of the file to be loaded. The actual loading is done in the `run` method using the `read_h5py_all_dataset` function.